[17.0][OU-FIX] base_automation: take care reviews
* move m2o_to_x2m in post migration
* move action server base automation name as end script
* rotate webhook uuid
* create trg_field_ref/trg_selection_field_id at pre-migration step
* revert _base_automation_update_deprecate_trigger to keep deprecated until next version
